Extract logic to base class
