+++ src/ILLink.CodeFix/BaseAttributeCodeFixProvider.cs
@@ -1,0 +1,92 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeActions;
+using Microsoft.CodeAnalysis.CodeFixes;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.Editing;
+using Microsoft.CodeAnalysis.Simplification;
+
+namespace ILLink.CodeFix
+{
+	public abstract class BaseAttributeCodeFixProvider : CodeFixProvider
+	{
+		public sealed override FixAllProvider GetFixAllProvider ()
+		{
+			// See https://github.com/dotnet/roslyn/blob/master/docs/analyzers/FixAllProvider.md for more information on Fix All Providers
+			return WellKnownFixAllProviders.BatchFixer;
+		}
+
+		internal async Task BaseRegisterCodeFixesAsync (CodeFixContext context, CodeFixProviderOperations.AttributeableParentTargets targets, string fullyQualifiedAttributeName, string s_title)
+		{
+			var root = await context.Document.GetSyntaxRootAsync (context.CancellationToken).ConfigureAwait (false);
+
+			var diagnostic = context.Diagnostics.First ();
+			var diagnosticSpan = diagnostic.Location.SourceSpan;
+
+			SyntaxNode targetNode = root!.FindNode (diagnosticSpan);
+			CSharpSyntaxNode? declarationSyntax = CodeFixProviderOperations.FindAttributableParent (targetNode, targets);
+			if (declarationSyntax is not null) {
+				var semanticModel = await context.Document.GetSemanticModelAsync (context.CancellationToken).ConfigureAwait (false);
+				var symbol = semanticModel!.Compilation.GetTypeByMetadataName (fullyQualifiedAttributeName);
+				var document = context.Document;
+				var editor = new SyntaxEditor (root, document.Project.Solution.Workspace);
+				var generator = editor.Generator;
+				var attrArgs = SetAttributeArguments (semanticModel, targetNode, declarationSyntax, generator, diagnostic);
+
+				// Register a code action that will invoke the fix.
+				context.RegisterCodeFix (
+					CodeAction.Create (
+						title: s_title,
+						createChangedDocument: c => AddAttribute (
+							document, editor, generator, declarationSyntax, attrArgs, symbol!, c),
+						equivalenceKey: s_title),
+					diagnostic);
+			}
+		}
+
+		private static async Task<Document> AddAttribute (
+			Document document,
+			SyntaxEditor editor,
+			SyntaxGenerator generator,
+			CSharpSyntaxNode containingDecl,
+			SyntaxNode[] attrArgs,
+			ITypeSymbol AttributeSymbol,
+			CancellationToken cancellationToken)
+		{
+			var semanticModel = await document.GetSemanticModelAsync (cancellationToken).ConfigureAwait (false);
+			if (semanticModel is null) {
+				return document;
+			}
+			var newAttribute = generator
+				.Attribute (generator.TypeExpression (AttributeSymbol), attrArgs)
+				.WithAdditionalAnnotations (
+					Simplifier.Annotation,
+					Simplifier.AddImportsAnnotation);
+
+			editor.AddAttribute (containingDecl, newAttribute);
+
+			return document.WithSyntaxRoot (editor.GetChangedRoot ());
+		}
+
+		internal abstract SyntaxNode[] SetAttributeArguments (SemanticModel semanticModel, SyntaxNode targetNode, CSharpSyntaxNode declarationSyntax, SyntaxGenerator generator, Diagnostic diagnostic);
+
+		internal static bool HasPublicAccessibility (IMethodSymbol? m)
+		{
+			if (m is not { DeclaredAccessibility: Accessibility.Public or Accessibility.Protected }) {
+				return false;
+			}
+			for (var t = m.ContainingType; t is not null; t = t.ContainingType) {
+				if (t.DeclaredAccessibility != Accessibility.Public) {
+					return false;
+				}
+			}
+			return true;
+		}
+	}
+}
