+++ src/ILLink.CodeFix/UnconditionalSuppressMessageCodeFixProvider.cs
@@ -5,21 +5,17 @@
 using System.Collections.Immutable;
 using System.Composition;
 using System.Globalization;
-using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using ILLink.RoslynAnalyzer;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Editing;
-using Microsoft.CodeAnalysis.Simplification;
 
 namespace ILLink.CodeFix
 {
 	[ExportCodeFixProvider (LanguageNames.CSharp, Name = nameof (UnconditionalSuppressMessageCodeFixProvider)), Shared]
-	public class UnconditionalSuppressMessageCodeFixProvider : CodeFixProvider
+	public class UnconditionalSuppressMessageCodeFixProvider : BaseAttributeCodeFixProvider
 	{
 		private const string s_title = "Add UnconditionalSuppressMessage attribute to parent method";
 		const string UnconditionalSuppressMessageAttribute = nameof (UnconditionalSuppressMessageAttribute);
@@ -28,53 +24,13 @@ namespace ILLink.CodeFix
 		public sealed override ImmutableArray<string> FixableDiagnosticIds
 			=> ImmutableArray.Create (RequiresUnreferencedCodeAnalyzer.DiagnosticId, RequiresAssemblyFilesAnalyzer.IL3000, RequiresAssemblyFilesAnalyzer.IL3001, RequiresAssemblyFilesAnalyzer.IL3002);
 
-		public sealed override FixAllProvider GetFixAllProvider ()
-		{
-			// See https://github.com/dotnet/roslyn/blob/master/docs/analyzers/FixAllProvider.md for more information on Fix All Providers
-			return WellKnownFixAllProviders.BatchFixer;
-		}
-
 		public sealed override async Task RegisterCodeFixesAsync (CodeFixContext context)
 		{
-			var root = await context.Document.GetSyntaxRootAsync (context.CancellationToken).ConfigureAwait (false);
-
-			var diagnostic = context.Diagnostics.First ();
-			var diagnosticSpan = diagnostic.Location.SourceSpan;
-
-			SyntaxNode targetNode = root!.FindNode (diagnosticSpan);
-			CSharpSyntaxNode? declarationSyntax = CodeFixProviderOperations.FindAttributableParent (targetNode, CodeFixProviderOperations.AttributeableParentTargets.All);
-
-			if (declarationSyntax is not null) {
-				var semanticModel = await context.Document
-					.GetSemanticModelAsync (context.CancellationToken).ConfigureAwait (false);
-				var symbol = semanticModel!.Compilation.GetTypeByMetadataName (FullyQualifiedUnconditionalSuppressMessageAttribute);
-
-				// Register a code action that will invoke the fix.
-				context.RegisterCodeFix (
-					CodeAction.Create (
-						title: s_title,
-						createChangedDocument: c => AddUnconditionalSuppressMessage (
-							context.Document, root, declarationSyntax, diagnostic, symbol!, c),
-						equivalenceKey: s_title),
-					diagnostic);
-			}
+			await BaseRegisterCodeFixesAsync (context, CodeFixProviderOperations.AttributeableParentTargets.All, FullyQualifiedUnconditionalSuppressMessageAttribute, s_title);
 		}
-		private static async Task<Document> AddUnconditionalSuppressMessage (
-			Document document,
-			SyntaxNode root,
-			CSharpSyntaxNode containingDecl,
-			Diagnostic diagnostic,
-			ITypeSymbol UnconditionalSuppressMessageSymbol,
-			CancellationToken cancellationToken)
+
+		internal override SyntaxNode[] SetAttributeArguments (SemanticModel semanticModel, SyntaxNode targetNode, CSharpSyntaxNode containingDecl, SyntaxGenerator generator, Diagnostic diagnostic)
 		{
-			var editor = new SyntaxEditor (root, document.Project.Solution.Workspace);
-			var generator = editor.Generator;
-
-			var semanticModel = await document.GetSemanticModelAsync (cancellationToken).ConfigureAwait (false);
-			if (semanticModel is null) {
-				return document;
-			}
-
 			// UnconditionalSuppressMessage("Rule Category", "Rule Id", Justification = "<Pending>")
 			var category = generator.LiteralExpression (diagnostic.Descriptor.Category);
 			var categoryArgument = generator.AttributeArgument (category);
@@ -87,17 +43,8 @@ namespace ILLink.CodeFix
 			var justificationExpr = generator.LiteralExpression ("<Pending>");
 			var justificationArgument = generator.AttributeArgument ("Justification", justificationExpr);
 
-			SyntaxNode[] attrArgs = new[] { categoryArgument, ruleIdArgument, justificationArgument };
-
-			var newAttribute = generator
-				.Attribute (generator.TypeExpression (UnconditionalSuppressMessageSymbol), attrArgs)
-				.WithAdditionalAnnotations (
-					Simplifier.Annotation,
-					Simplifier.AddImportsAnnotation);
-
-			editor.AddAttribute (containingDecl, newAttribute);
-
-			return document.WithSyntaxRoot (editor.GetChangedRoot ());
+			return new[] { categoryArgument, ruleIdArgument, justificationArgument };
 		}
+
 	}
 }
